Aggregate data with start and end dates for specific signals
These are edits necessary to conduct the ATSPM analysis that BYU conducted in 2019-2020. As a note, the UDOT ATSPM project continued between when we began and when we are making this push, so a number of changes appear that are not material to the work we accomplished. Further cleaning would be necessary to make a clean pull request if that is desired.